Change whoGoesFirst to changeSymbol function
Comment out the whoGoesFirst function and replace it with
changeSymbol, a function that switches back and forth between the
user being X and O every other game with the computer.  The user now
always goes first as X, and then the computer will go first as X the
following game, and so on.  This makes more sense seeing that the
player going first has an advantage.